pyproject.toml
[project]
name = "sd-marag"
version = "0.1.0"
description = "Self-Decisive Multi-Agentic Retrieval Augmented Generation"
authors = [
    {name = "Praneet Pabolu",email = "pp542@cornell.edu"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "pymilvus (>=2.5.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.59.7,<2.0.0)",
    "datasets (>=3.2.0,<4.0.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "transformers (>=4.48.0,<5.0.0)",
    "accelerate (>=1.3.0,<2.0.0)",
    "protobuf (>=5.29.3,<6.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "sentence-transformers (>=3.3.1,<4.0.0)",
    "pymupdf (>=1.25.2,<2.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "vllm (>=0.7.0,<0.8.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "markdown-pdf (>=1.6,<2.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pymongo (>=4.12.0,<5.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pdfnumbering (>=0.1.1,<0.2.0)",
    "fpdf2 (==2.8.3)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
